New approach for insertion.
Consists of a lock handler using the OpenMP omp_lock_t. Also, the values are split among the threads which add them in nondecreasing order
(accepts repeated entries) to the main list. Threads do not have a private list in this case.

Small benchs have demonstrated speed down. However, the purpose for now is test the lock handler.

Next steps are:
- Test the locking strategy with priorities.
- Find what is causing the slowness, look up for possible bottlenecks.