Locking with priority working.
Some insertions have shown that locking with priority in this strategy has not ended up in deadlocks, both for increasing and nonincreasing prioritites. As increasing the priority is done in critical section, no race condition was found.

Running time is a way slower than sequential. Nonetheless, discovering that lock handler does not have deadlocks and data reace is an important fact for now.